Added implementation files folder (src), added figures.dart file where export all figures
