+++ lib/figures.dart
@@ -1,0 +1,7 @@
+export 'src/ballena.dart' show Ballena;
+export 'src/caballito.dart' show Caballito;
+export 'src/caballo.dart' show Caballo;
+export 'src/iguana.dart' show Iguana;
+export 'src/mochila.dart' show Mochila;
+export 'src/pinguino.dart' show Pinguino;
+export 'src/stickman.dart' show Stickman;
